requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.1.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
Werkzeug==3.0.1
